+++ hata/discord/integration/preinstanced.py
@@ -1,11 +1,11 @@
-__all__ = ('IntegrationExpireBehavior',)
+__all__ = ('IntegrationExpireBehavior', 'IntegrationType')
 
 from ..bases import Preinstance as P, PreinstancedBase
 
 
 class IntegrationExpireBehavior(PreinstancedBase):
     """
-    Represents an ``ApplicationDetail``'s expire behavior.
+    Represents an ``IntegrationDetail``'s expire behavior.
     
     Attributes
     ----------
@@ -23,7 +23,7 @@ class IntegrationExpireBehavior(PreinstancedBase):
     DEFAULT_NAME : `str` = `'Undefined'`
         The default name of the integration expire behaviors.
     
-    Every predefined message activity type can be accessed as class attribute as well:
+    Every predefined expire behavior can be accessed as class attribute as well:
     
     +-----------------------+---------------+-------+
     | Class attribute name  | name          | value |
@@ -42,3 +42,68 @@ class IntegrationExpireBehavior(PreinstancedBase):
     remove_role = P(0, 'remove_role')
     kick = P(1, 'kick')
 
+
+
+class IntegrationType(PreinstancedBase):
+    """
+    Represents an ``Integration``'s type.
+    
+    Attributes
+    ----------
+    name : `str`
+        The name of the integration type.
+    value : `int`
+        The Discord side identifier value of the integration type.
+    
+    Class Attributes
+    ----------------
+    INSTANCES : `dict` of (`int`, ``IntegrationType``) items
+        Stores the predefined ``IntegrationType``-s. These can be accessed with their `value` as key.
+    VALUE_TYPE : `type` = `int`
+        The integration type' values' type.
+    DEFAULT_NAME : `str` = `'Undefined'`
+        The default name of the integration types.
+    
+    Every predefined integration type can be accessed as class attribute as well:
+    
+    +-----------------------+---------------+---------------+
+    | Class attribute name  | name          | value         |
+    +=======================+===============+===============+
+    | none                  | none          | `''`          |
+    +-----------------------+---------------+---------------+
+    | discord               | discord       | `'discord'`   |
+    +-----------------------+---------------+---------------+
+    | youtube               | youtube       | `'youtube'`   |
+    +-----------------------+---------------+---------------+
+    """
+    INSTANCES = {}
+    VALUE_TYPE = str
+    
+    __slots__ = ()
+    
+
+    @classmethod
+    def _from_value(cls, value):
+        """
+        Creates a new integration type with the given value.
+        
+        Parameters
+        ----------
+        value : `str`
+            The integration's type.
+        
+        Returns
+        -------
+        self : ``IntegrationType``
+            The created instance.
+        """
+        self = object.__new__(cls)
+        self.name = value
+        self.value = value
+        
+        return self
+    
+    # predefined
+    none = P('', 'none')
+    discord = P('discord', 'discord')
+    youtube = P('youtube', 'youtube')