+++ changelog.md
@@ -4,10 +4,12 @@
 
 - Add `NsfwLevel.nsfw`.
 - `Message.guild` now works for invoking user only messages as well. (Gilgamesh#8939)
+- Add `IntegrationType`.
 
 #### Bug Fixes
 
 - `Guild.nsfw` returned the opposite value. (Gilgamesh#8939)
+- `IntegrationAccount.__new__` could raise `TypeError`. (since 1.3.4)
 
 ## 1.3.5 *\[2022-08-27\]*